refactor(catalog): split api/helpers.py into images and rich_text modules
Move image/media extraction helpers into apps/catalog/api/images.py and
build_rich_text into apps/catalog/api/rich_text.py. Drop lying underscores
from helpers shared across sibling modules (serialize_credit, serialize_title_ref,
serialize_title_machine, collect_titles, serialize_locations).

Behavior-preserving — no schema or response changes.
