+++ backend/apps/catalog/api/images.py
@@ -1,0 +1,202 @@
+"""Image and uploaded-media helpers for catalog API endpoints."""
+
+from __future__ import annotations
+
+from collections.abc import Iterable, Sequence
+from typing import Any
+
+from django.db.models import Prefetch
+
+from apps.core.licensing import (
+    UNKNOWN_LICENSE_RANK,
+    get_minimum_display_rank,
+)
+from apps.core.types import JsonData
+from apps.media.models import EntityMedia
+from apps.media.schemas import MediaRenditionsSchema, UploadedMediaSchema
+from apps.media.storage import build_public_url, build_storage_key
+from apps.provenance.schemas import AttributionSchema
+
+from ..models import CorporateEntity
+
+__all__ = [
+    "extract_image_attribution",
+    "extract_image_urls",
+    "first_thumbnail",
+    "media_prefetch",
+    "serialize_uploaded_media",
+]
+
+
+# Slot 2 is Any because prefetch_related has a single _PrefetchedQuerySetT
+# TypeVar it must unify across all heterogeneous Prefetch args at the call
+# site; any concrete queryset type here breaks that unification. The Any
+# is an artifact of django-stubs' API design, not lost information.
+def media_prefetch() -> Prefetch[str, Any, str]:
+    """Return a Prefetch for ready EntityMedia with assets."""
+    return Prefetch(
+        "entity_media",
+        queryset=EntityMedia.objects.filter(
+            asset__status="ready",
+        ).select_related("asset", "asset__uploaded_by"),
+        to_attr="all_media",
+    )
+
+
+def serialize_uploaded_media(
+    all_media: Iterable[EntityMedia],
+) -> list[UploadedMediaSchema]:
+    """Serialize EntityMedia rows into the uploaded_media response list."""
+    return [
+        UploadedMediaSchema(
+            asset_uuid=str(em.asset.uuid),
+            category=em.category,
+            is_primary=em.is_primary,
+            uploaded_by_username=(
+                em.asset.uploaded_by.username if em.asset.uploaded_by else None
+            ),
+            renditions=MediaRenditionsSchema(
+                thumb=build_public_url(build_storage_key(em.asset.uuid, "thumb")),
+                display=build_public_url(build_storage_key(em.asset.uuid, "display")),
+            ),
+        )
+        for em in all_media
+    ]
+
+
+def _uploaded_image_urls(
+    primary_media: Sequence[EntityMedia] | None,
+) -> tuple[str | None, str | None]:
+    """Return (thumbnail_url, hero_image_url) from prefetched EntityMedia.
+
+    Prefers ``backglass`` category, then falls back to any primary.
+    Returns ``(None, None)`` when no uploaded media is available.
+    """
+    if not primary_media:
+        return None, None
+
+    # Prefer backglass, fall back to first available primary.
+    chosen = None
+    for em in primary_media:
+        if em.category == "backglass":
+            chosen = em
+            break
+    if chosen is None:
+        chosen = primary_media[0]
+
+    asset_uuid = chosen.asset.uuid
+    thumb = build_public_url(build_storage_key(asset_uuid, "thumb"))
+    hero = build_public_url(build_storage_key(asset_uuid, "display"))
+    return thumb, hero
+
+
+def extract_image_urls(
+    extra_data: JsonData,
+    primary_media: Sequence[EntityMedia] | None = None,
+    *,
+    min_rank: int | None = None,
+) -> tuple[str | None, str | None]:
+    """Return (thumbnail_url, hero_image_url).
+
+    When *primary_media* (prefetched ``EntityMedia`` rows) contains uploaded
+    images, those are used unconditionally (no license gating — Pinbase owns
+    them).  Otherwise falls back to third-party images in *extra_data*,
+    respecting the global Constance display threshold.
+
+    Pass *min_rank* to avoid repeated Constance DB lookups in tight loops.
+    """
+    # Uploaded media always wins — no license gating.
+    thumb, hero = _uploaded_image_urls(primary_media)
+    if thumb or hero:
+        return thumb, hero
+
+    if min_rank is None:
+        min_rank = get_minimum_display_rank()
+
+    def _rank_ok(key: str) -> bool:
+        rank = extra_data.get(f"{key}.__permissiveness_rank")
+        effective = rank if isinstance(rank, int) else UNKNOWN_LICENSE_RANK
+        return effective >= min_rank
+
+    def _abs(url: str | None) -> str | None:
+        """Return *url* only if it's an absolute HTTP(S) URL, else None."""
+        if url and (url.startswith("http://") or url.startswith("https://")):
+            return url
+        return None
+
+    # Try OPDB structured images first (have size variants).
+    images = extra_data.get("opdb.images")
+    if images and isinstance(images, list) and _rank_ok("opdb.images"):
+        img = None
+        for candidate in images:
+            if isinstance(candidate, dict) and candidate.get("primary"):
+                img = candidate
+                break
+        if img is None:
+            img = images[0] if images else None
+        if isinstance(img, dict):
+            urls = img.get("urls") or {}
+            thumbnail = _abs(urls.get("medium") or urls.get("small"))
+            hero = _abs(urls.get("large") or urls.get("medium"))
+            if thumbnail or hero:
+                return thumbnail, hero
+
+    # Fall back to flat URL list (IPDB-sourced or scraped).
+    for key in ("ipdb.image_urls", "image_urls"):
+        image_urls = extra_data.get(key)
+        if image_urls and isinstance(image_urls, list) and _rank_ok(key):
+            first = image_urls[0]
+            if isinstance(first, str) and _abs(first):
+                return first, first
+
+    return None, None
+
+
+def extract_image_attribution(
+    extra_data: JsonData,
+    primary_media: Sequence[EntityMedia] | None = None,
+) -> AttributionSchema | None:
+    """Return AttributionSchema for the displayed image, or None.
+
+    When uploaded media is being used (determined by *primary_media*), returns
+    ``None`` — no third-party license to cite.  Otherwise checks each external
+    image source in priority order and returns info for the first source that
+    passes the display threshold.
+    """
+    # Uploaded media has no third-party attribution.
+    if primary_media:
+        return None
+
+    min_rank = get_minimum_display_rank()
+
+    for key in ("opdb.images", "ipdb.image_urls", "image_urls"):
+        data = extra_data.get(key)
+        if not data:
+            continue
+        rank_raw = extra_data.get(f"{key}.__permissiveness_rank")
+        rank = rank_raw if isinstance(rank_raw, int) else None
+        effective = rank if rank is not None else UNKNOWN_LICENSE_RANK
+        if effective >= min_rank:
+            license_slug_raw = extra_data.get(f"{key}.__license_slug")
+            license_slug = (
+                license_slug_raw if isinstance(license_slug_raw, str) else None
+            )
+            return AttributionSchema(
+                license_slug=license_slug,
+                permissiveness_rank=rank,
+            )
+
+    return None
+
+
+def first_thumbnail(
+    entities_with_models: Iterable[CorporateEntity], *, min_rank: int
+) -> str | None:
+    """Return the first non-None thumbnail URL from nested entity→model prefetches."""
+    for entity in entities_with_models:
+        for model in entity.models.all():
+            if model.extra_data:
+                thumb, _ = extract_image_urls(model.extra_data, min_rank=min_rank)
+                if thumb:
+                    return thumb
+    return None
