+++ backend/apps/catalog/tests/test_image_licensing.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from apps.catalog.api.helpers import _extract_image_urls
+from apps.catalog.api.images import extract_image_urls
 from apps.catalog.models import Title
 from apps.catalog.resolve import resolve_model
 from apps.catalog.tests.conftest import make_machine_model
@@ -185,7 +185,7 @@ class TestExtractImageUrlsWithThreshold:
             "opdb.images.__license_slug": "not-allowed",
             "opdb.images.__permissiveness_rank": 0,
         }
-        thumb, hero = _extract_image_urls(extra_data)
+        thumb, hero = extract_image_urls(extra_data)
         assert thumb is None
         assert hero is None
 
@@ -205,7 +205,7 @@ class TestExtractImageUrlsWithThreshold:
             "ipdb.image_urls": ["https://ipdb.org/ipdb.jpg"],
             "ipdb.image_urls.__permissiveness_rank": 85,  # above threshold
         }
-        thumb, _ = _extract_image_urls(extra_data)
+        thumb, _ = extract_image_urls(extra_data)
         assert thumb == "https://ipdb.org/ipdb.jpg"
 
     def test_null_rank_uses_unknown_rank(self):
@@ -223,7 +223,7 @@ class TestExtractImageUrlsWithThreshold:
             "opdb.images.__permissiveness_rank": None,  # unknown
         }
         # With "licensed-only" (min_rank=38), unknown (rank 5) should be hidden.
-        thumb, hero = _extract_image_urls(extra_data)
+        thumb, hero = extract_image_urls(extra_data)
         assert thumb is None
         assert hero is None
 
@@ -244,7 +244,7 @@ class TestExtractImageUrlsWithThreshold:
             "opdb.images.__permissiveness_rank": 0,
         }
         with override_config(CONTENT_DISPLAY_POLICY="show-all"):
-            thumb, _ = _extract_image_urls(extra_data)
+            thumb, _ = extract_image_urls(extra_data)
         assert thumb == "https://img.opdb.org/m.jpg"
 
     def test_include_unknown_shows_null_rank(self):
@@ -264,5 +264,5 @@ class TestExtractImageUrlsWithThreshold:
             "opdb.images.__permissiveness_rank": None,
         }
         with override_config(CONTENT_DISPLAY_POLICY="include-unknown"):
-            thumb, _ = _extract_image_urls(extra_data)
+            thumb, _ = extract_image_urls(extra_data)
         assert thumb == "https://img.opdb.org/m.jpg"
